+++ backend/routers/auth.py
@@ -22,14 +22,13 @@ def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
 def authenticate_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
     """Check correct authentication"""
     email = user.email
-    hashed_pwd = get_hashed_password(user.password)
     
-
     db_user = db.query(models.User).filter(models.User.email == email).first()
     if db_user is None:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"The user {email} does not exist")
     
-    if db_user.hashed_password != hashed_pwd:
+    if not verify_password(user.password, db_user.hashed_password):
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"The credentials you entered are not correct")
+
 
     return db_user