+++ frontend/src/Components/AdminView.js
@@ -1,38 +1,50 @@
+import { isAxiosError } from "axios"
 import { useAxiosGet } from "../Hooks/HttpRequestsAxios"
 
 function AdminView() {
 
     const adminUrl = "http://localhost:8000/admin/users"
     let usersRendered
-    let tableRendered
+    let tableRendered = <div></div>
+    let responseResult
     let len = "0"
 
-    const userList = useAxiosGet(adminUrl).data
-    console.log(typeof (userList))
+    const userListResponse = useAxiosGet(adminUrl).data
+    console.log(userListResponse)
 
-    if (userList) {
-        usersRendered = userList.map(d => (<tr key = {d.id}><td className="border-2">{d.id}</td><td className="border-2">{d.email}</td><td className="border-2">{d.hashed_password}</td></tr>))
-        tableRendered = (<table className="border-2 table-auto">
-            <thead>
-                <tr>
-                <th>ID</th>
-                <th>Username</th>
-                <th>Hashed password</th>
-                </tr>
-            </thead>
-            <tbody>
-            {usersRendered}
-            </tbody>
-        </table>)
-        len = Object.keys(userList).length
+    if (userListResponse) {
+        if (isAxiosError(userListResponse)) {
+
+            console.log("message to print", userListResponse.message)
+            responseResult = <div> {userListResponse.message}: There seems to be an issue with retrieving the requested information. Check the server works as expected.</div>
+        }
+        else {
+            usersRendered = userListResponse.map(d => (<tr key={d.id}><td className="border-2">{d.id}</td><td className="border-2">{d.email}</td><td className="border-2">{d.hashed_password}</td></tr>))
+            tableRendered = (<table className="border-2 table-auto">
+                <thead>
+                    <tr>
+                        <th>ID</th>
+                        <th>Username</th>
+                        <th>Hashed password</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {usersRendered}
+                </tbody>
+            </table>)
+            len = Object.keys(userListResponse).length
+            responseResult = <div>There are currently {len} users</div>
+        }
+
     }
 
 
 
     return (
-        <h1> There are currently {len} users
+        <div>
+            {responseResult}
             {tableRendered}
-        </h1>
+        </div>
     )
 }