+++ frontend/src/Components/Navigation.js
@@ -1,20 +1,22 @@
-import React, { useState , useEffect} from "react";
+import React, { useState, useEffect } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faBars } from "@fortawesome/free-solid-svg-icons"
 import { Link } from "react-router-dom";
-import Menu  from "./Menu";
+import Menu from "./Menu";
 
 function Navigation() {
 
     //react hook to know if menu is open or not
     const [showMenu, setShowMenu] = useState(false)
-    const closeMenu = ()=>setShowMenu(false)
+    const closeMenu = () => setShowMenu(false)
     let menu
 
-    //conditional rendering
+    //useEffect(() => { window.addEventListener("click", closeMenu) }, []); // [] so that the hook runs when the component first renders
 
+
+    //conditional rendering
     if (showMenu) {
-        menu = <Menu closeMenu = {closeMenu} />
+        menu = <Menu closeMenu={closeMenu} />
     }
     return (
         <nav>